+++ dimcli/tests/test_dataframes.py
@@ -96,6 +96,15 @@ class TestDataframes(unittest.TestCase):
         # ----
         click.secho("Completed test succesfully", fg="green")
 
+    def test_005(self):
+        click.secho("\nTEST 005: Concepts extraction.", fg="green")
+        # ----
+        print("Testing as_dataframe_funders on Grants data: ")
+        res= dslquery("""search publications for "graphene" where year=2019 return publications[id+concepts+year+title] limit 100""")
+        concepts = res.as_dataframe_concepts()
+        print(" ==> res.as_dataframe_concepts(): ", res.as_dataframe_concepts())
+        print(concepts.info())
+        click.secho("Completed test succesfully", fg="green")
 
 if __name__ == "__main__":
     unittest.main()